% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert package data to various formats}
\usage{
convert(input, name, format, output)
}
\arguments{
\item{input}{filename of input data. Can be lazyload datbase.}

\item{name}{data name to import}

\item{format}{what format to export to}

\item{output}{filename to write output to}
}
\description{
Wrapper for r-universe data export tool. The data import functions are copied
from utils::data(), reading supoprted types and also lazydata databases.
}
