% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_boolean.R
\name{is_boolean}
\alias{is_boolean}
\alias{is_TRUE}
\alias{is_FALSE}
\title{Identify Boolean Values}
\usage{
is_boolean(x)

is_TRUE(x)

is_FALSE(x)
}
\arguments{
\item{x}{An atomic or recursive vector (i.e., c(), matrix(), list(), or data.frame()).}
}
\value{
A logical vector which has the same length and dimensions with \code{x}.
}
\description{
A wrapper for the base::is.logical(), base::isTRUE(), and base::isFALSE().
}
\examples{
test <- list(
    num = c(1.2, -3.3, 7.5, 2.9),
    char = c(LETT = LETTERS[1:3], lett = letters[1:3]),
    cplx = matrix(complex(real = 1:12, imaginary = 3:9), nrow = 3),
    bool = data.frame(true = rep(TRUE, 5), false = rep(FALSE, 5)),
    na = list(NA, NA_character_, NA_complex_, NA_integer_, NA_real_),
    special = list(TRUE, FALSE, NaN, NULL)
)
is_TRUE(test)
is_FALSE(test)
is_boolean(test)
}
