% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_character.R
\name{is_character}
\alias{is_character}
\title{Identify Character Values}
\usage{
is_character(x)
}
\arguments{
\item{x}{An atomic or recursive vector (i.e., c(), matrix(), list(), or data.frame()).}
}
\value{
A logical vector which has the same length / dimensions with \code{x}.
}
\description{
A wrapper for the base::is.character().
}
\examples{
test <- list(
    num = c(1.2, -3.3, 7.5, 2.9),
    char = list(LETT = LETTERS[1:3], lett = letters[1:3]),
    cplx = matrix(complex(real = 1:12, imaginary = 3:9), nrow = 3),
    bool = data.frame(true = rep(TRUE, 5), false = rep(FALSE, 5)),
    na = list(NA, NA_character_, NA_complex_, NA_integer_, NA_real_),
    special = list(TRUE, FALSE, NaN, NULL)
)
is_character(test)
}
