% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_reshape.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Reshape a data frame from wide format to long format}
\usage{
wide_to_long(
  data,
  cols,
  keys_name = character(1),
  vals_name = character(1),
  keep_unused = TRUE
)
}
\arguments{
\item{data}{A data.frame input}

\item{cols}{A character/numeric vector of column names or indices to select the columns.}

\item{keys_name}{A new column name for the \code{cols}.}

\item{vals_name}{A new column name for the values from the \code{cols}.}

\item{keep_unused}{Logical (default: TRUE). Keep the other columns that are not included in \code{cols}.}
}
\value{
A reshaped data.frame
}
\description{
Reshape a data frame from wide format to long format
}
\examples{
set.seed(1)
df0 <- as.data.frame(matrix(1:25, ncol = 5, dimnames = list(1:5, LETTERS[1:5])))
wide_to_long(df0, cols = c("A", "B", "D"), keys_name = "group", vals_name = "length")
wide_to_long(df0, cols = c(2, 3, 5), keys_name = "group", vals_name = "length", keep_unused = FALSE)
}
