% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integer.R
\name{is_integer}
\alias{is_integer}
\title{Identify Integer Values}
\usage{
is_integer(x)
}
\arguments{
\item{x}{An atomic or recursive vector (i.e., c(), matrix(), list(), or data.frame()).}
}
\value{
A logical vector which has the same length and dimensions with \code{x}.
}
\description{
A wrapper for the base::is.integer().
}
\examples{
test <- list(
    num = c(1.2, -3.3, 7.5, 2.9),
    int = 1:5,
    char = c(LETT = LETTERS[1:3], lett = letters[1:3]),
    cplx = matrix(complex(real = 1:12, imaginary = 3:9), nrow = 3),
    bool = data.frame(true = rep(TRUE, 5), false = rep(FALSE, 5)),
    na = list(NA, NA_character_, NA_complex_, na_int = NA_integer_, NA_real_),
    special = list(TRUE, FALSE, NaN, NULL)
)
is_integer(test)
}
